Fix bad UX by replacing iView grid with flexbox
The draggable area is now the same height for each column. When
dragging an item into an empty column, you don't have to hover to
the small space at the top. Also the fixed text size was causing a
visual issue on mobile screens so I used viewport width to scale the
text.